refactor _app 코드 정리 + 유저 로직 완료
