+++ src/store/index.ts
@@ -1,18 +1,20 @@
-import { configureStore, applyMiddleware } from '@reduxjs/toolkit';
+import { configureStore, applyMiddleware, combineReducers } from '@reduxjs/toolkit';
 import createSagaMiddleware from 'redux-saga';
 import auth from './reducer/auth';
 import asider from './reducer/asider';
 import darkMode from './reducer/darkmode';
 import rootSaga from './sagas';
 
+const reducer = combineReducers({
+  auth,
+  asider,
+  darkMode,
+});
+
 const sageMiddleware = createSagaMiddleware();
 
 const store = configureStore({
-  reducer: {
-    auth,
-    asider,
-    darkMode,
-  },
+  reducer,
   middleware: [sageMiddleware],
   enhancers: [applyMiddleware(sageMiddleware)],
 });
